+++ spec/models/dorsale/customer_vault/event_spec.rb
@@ -5,10 +5,10 @@ RSpec.describe ::Dorsale::CustomerVault::Event, type: :model do
   it { is_expected.to belong_to :person }
   it { is_expected.to belong_to :comment }
 
-  it { is_expected.to validate_presence_of :author }
   it { is_expected.to validate_presence_of :person }
   it { is_expected.to validate_presence_of :action }
 
+  it { is_expected.to_not validate_presence_of :author }
   it { is_expected.to_not validate_presence_of :comment }
 
   it "should validate comment if action is comment" do