+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Change nilify_blanks config
 - CustomerVault Events : persist text
+- CustomerVault : delete event when deleting comment
 
 ## 3.7.3
 - Fix copy action